documentation in src/rootfile.scala and src/dkcheck.scala + wip
